+++ GANDLF/training_manager.py
@@ -302,6 +302,10 @@ def TrainingManager(dataframe, outputDir, parameters, device, resume, reset):
         if singleFoldTesting:
             break
         currentTestingFold += 1  # go to next fold
+    
+    # End by copying all results, if requested
+    if parameters["second_output_dir"]:
+        shutil.copytree(Path(outputDir), Path(parameters["second_output_dir"]), dirs_exist_ok=True)
 
 
 def TrainingManager_split(